Snackbar now detects if it should offset by the tab bar
Typically, the snackbar should be shown with a margin to the bottom tab
bar. However, in some flows, such as the Receive address copy flow, the
snackbar is triggered when there is no tab bar visible. In these cases,
the snackbar should be shown without that offset.

Because the snackbar is a top-level component, and the routes and ways
that nested components may or may not hide the tab bar shouldn't be
prescribed by the tab bar, the snackbar now does a hit test to see if
the element at the center point of the tab bar area is the tab bar or
something else. If no tab bar is found, the offset is adjusted to ignore
the tab bar.

The outcome is somewhat hacky, but works quite well.