Fix a useEffect/setState loop in NetworkSettingsSelect
At initial NetworkSettingsSelect render, and any time gas options and/or
active fee index are updated, the component notifies the parent of a
change in the network settings. The way this was set up, any change in
the networkSettings prop would trigger an attempt to update the parent,
but if the parent used onNetworkSettingsChange to update that prop
(which is a normal way to set things up), a render loop would ensue.

The reason the useEffect depended on networkSettings as a whole was
solely so that it could preserve the current gas limit unchanged when
invoking `onNetworkSettingsChange`. Changing the dependency array for
the responsible `onNetworkSettingsChange`-triggering `useEffect` to
specifically include `networkSettings.gasLimit` instead of
`networkSettings` as a whole takes care of this issue.

Misleadingly, this manifested as a render loop in the
`EstimateRefreshCountdownDivider` component---it's unclear why this was
the case, but fixing the above removed the
`EstimateRefreshCountdownDivider` render loop.