extend emittery so `emit()` returns unknown[] The return value from all the listeners if any.
We did NOT have a way to return temporary data from the background to the UI.
All the data from the background had to go through `dispatch > reducer > store > selector` path
even if the data was ephemeral eg. time bound quotes or single use ids

The `emitter.emit()` already returns a promise, so the infrastructure was in place.
The only problem was that it did not propagate/allow the return value from the listeners.
This patch makes this possible and updates the typings of the lib.

Yarn workspace has installed `emittery` into `node_modules` and `background/node_modules`.
When importing, code from `background/node_modules` was used.
To be able to patch the package it had to moved to `node_modules`.

When the `emittery` package was not present in the `background/package.json`
the `import/no-extraneous-dependencies` linter rule throw an error when importing emittery.
To solution for this is to inform the this rule that we have multiple `package.json` files,
and dependencies can be used from these.

Tried to keep the `emittery` package in the `background` only, but the patch-package
did not run on `yarn install` with that setup.