Drop isTypeDestination from SharedAssetInput
This mode was completely orthogonal to the other usage of
SharedAssetInput, allowing input of an arbitrary address associated with
no asset info. It was also only used in Send. The only commonality it
had with the rest of the component was styling.

Instead of having this unrelated code path in SharedAssetInput, do some
custom styling for the address input in Send.tsx.