Drop note to use existing conversion logic in Send
Currently the Send component does custom JS number conversions using the
unitPrice on the selected asset. Instead, it should be using asset price
points and convertAssetAmountViaPricePoint to make sure to be consistent
with the rest of the system.

Leave a FIXME in place to fix this when the time comes.