Add AssetAmount support to SharedAssetInput
SharedAssetInput now takes an assetsAndAmounts prop, which is of a new
type, AnyAssetWithOptionalAmount (defined in SharedAssetItem for now).
This type can contain either an asset, or an asset with associated
amount and localized decimal information.

This then cascades into a few additional bits of functionality:

- SharedAssetItem now displays the associated amount in the dropdown if
  available.
- SharedAssetInput now displays the associated amount for the selected
  asset (if any) as a balance.
- SharedAssetInput now displays, by default, a "Max" button to set the
  input amount to the associated amount for the selected asset.

Additionally, the SharedAssetInput now parses the inputted amount using
fixed-point parsing and compares the fixed-point amount to the max
balance using fixed-point comparisons instead of potentially-lossy
JS number parsing.