Rely on `forAccounts` to check account balances in IndexingService
Previously, it was hardcoded that only the `from` account balance would
be checked for an updated balance when a transaction cleared. This is
now done based on the `forAccounts` parameter passed with the
`transaction` event, which filters `to` and `from` accounts to see if
they are currently being tracked by the indexing service.

In the future, this could also be populated by the enrichment service
with additional accounts touched by the transaction that might require
balance updates, as well, making it a good future-facing approach to
filtering transactions for balance change interest.