Track transactions that failed to confirm
Currently, the only ones detected are transactions that failed to
broadcast in the first place; however, there will be space to also mark
those transactions that are broadcast but fall out of the mempool before
being confirmed.

Currently, these transactions are always stuck to the top of the
activity list; managing them is a separate concern.