Update local state through payloadCreator thunkAPI
Previously updating the state was not dispatched, with the use
of  thunkAPI the data is now saved. This commit also adds a
function that checks if the approval target is approved and
saved the approval contract and amount in redux state so when
user visits a specific vault page for the second time the allowance
call is not made again. This implies that the user will not go to
the contract and decrease allowance himself.