Fix an issue where gas limits did not pass through providers
Due to a difference between the JSON-RPC transaction fields and the ones
in EthersTransactionRequest, gas limits specified by callers of the
providers (both via the bridge and via the internal provider) were being
dropped/ignored internally. This is now fixed.

Additionally, there was a potential issue if the provider callers passed
the more correct and recent field `input` to carry contract data; in
these cases, there was no guarantee that it would be processed correctly
internally, as `EthersTransactionRequest` expects this data in the older
`data` field. This discrepancy is also now handled at the edge.

Both of these are fixed in `InternalEthereumProvider`, which ultimately
processes both internal and provider bridge transaction requests.