Add expiry to queued transaction lookup
If a transaction lookup fails for more than 10 hours after the
transaction's lookup is first requested, stop looking for it. If the
transaction in question had a corresponding entry in the database with
no status, set that status to failed.

This handles any historical pending transactions that may have been
dropped. Transactions that were confirmed should already have a status.
Additionally, future pending transactions that are not confirmed after
10 hours will be marked as failed.

Ideally there would be an "expired" status that is not specifically
failed, but that is not currently the case.