handle FF for ledger slice in selector
When the feature flag is off, this breaks after the #956 has been merged.

The break happens where the `state.ledger.devices`  is used inside a
`createSelector`, because `createSelector` is run an start up time
and not influenced by the fact that it's not reachable code.

The rest of usages of `state.ledger.devices` do not cause an issue
because they are used at runtime and if the code is not reacable it won't
throw an error.

Those instances should probably also be selectors though.