Don't clear out a pending transaction state
Transaction state is cleared when a new transaction is about to be
processed; in these cases, the transaction state is pending, but it was
being reset to idle. An idle transaction state leads the sign
transaction page to believe no transaction is available, and to exit
back to its caller.

Now, the transaction state is preserved if it is pending, otherwise it
is cleared to idle.