Rework selected asset handling in SharedAssetInput
Rather than have a defaultAsset, pass a selectedAsset, which is expected
to be the asset displayed as selected at all times. Consumers are
expected to manage the state of the currently selected asset with
onAssetSelect, as is done with most other form fields.

This also fixes the behavior of the flip swap button in swaps.