Don't clear selected fee type with transaction
The intent with selected fee type is to persist it from transaction to
transaction, rather than clear it after every transaction. This was
mostly implemented, but as calls to `clearTransactionState` became more
aggressive, it turned out that clearing the state also reset the fee
type. This adjustment fixes that, setting the fee type to “regular”
only if the previous state is empty.