Use Ethers effectiveGasPrice as post-confirmation gas price
Once a transaction has confirmed, the gasPrice field becomes the field
that represents the final used gas price on-chain.

For legacy transactions, this means we don't preserve information on the
gas price originally requested by the user; however, for EIP-1559
transactions that information is carried by maxFeePerGas and
maxPriorityFeePerGas.

This fixes an issue where the gas price would always equal the
maxFeePerGas post-confirmation.