Check that manifest version matches git tags
Current version is computed with `git describe --tags`
(option `--tags` allows for un-annotated tags).
The most recent tag is matched against the version in the manifest,
failing the build if a mismatch is found.
If the current commit is not exactly a tagged commit,
then the current version is not official and a warning is shown.