AddressNetwork->AddressOnNetwork, NameNetwork->NameOnNetwork
The original names weren't fully considered but mostly resulted from a
rename on AccountNetwork. `AddressNetwork` most directly implies the
address of a network (e.g. as with “domain name”, the name of a domain,
or “Tally doggo”, the doggo of Tally), while adding the "on" clarifies
that this type is meant to carry a particular address or name on a
particular network, as per the JSDocs.