Introduce ethereumNetwork instance variable in ChainService
This is a fairly transitive component of the transition away from
getEthereumNetwork, but reduces calls to the function and centralizes
responsibility for the network in ChainService, for now.

This commit introduces it as a purely internal tool, though it will
(temporarily) be consumed externally  soon. The main purpose here is to
make most references to ethereumNetwork or getEthereumNetwork simple
variable references to a single `network` local. This will then lend
itself to straightforward parameter refactoring.