Take network as parameter in signing routines
This is used when converting Ethers signed transactions to Tally
transaction data; most likely the correct long-term strategy here will
be to have the signing service take care of this conversion, and let
the signer services deal directly in Ethers data for simplicity. For
now, however, this parameter infects the whole call tree.