Use subannotations in indexing to look up asset balances
The indexing service is now notified when enriched transactions resolve,
and uses the subannotations to determine if it needs to fast-check
asset balances. This ensures that even assets that were not previously
being balance-checked but were involved in a transaction with an ERC20
transfer will be balance-checked.

Note that this will NOT balance-check tokens that are not in a token
list; for now, only known tokens can be detected this way.