Move signature request enrichment into updateOptions handler
This is temporary, but allows all signature requests to be enriched
before they are displayed. This prepares the wiring for having
transaction signing pages that edit the data in enrichment annotations
and must then be updated with updated annotations after; in the current
flow, signature requests are only enriched at the initial request,
removing the ability to update enrichment data from the signing page.