Extract decimal-for-unit-price heuristic to asset-utils
This utility is going to be used soon outside of the accounts selectors.

The utility takes an already-computed unit price as a JS number, but can
also be fed a UnitPricePoint. UnitPricePoint is now parametrized with
its asset type to clear up some return types as well.

Finally, the utility also deals with undefined unit prices by simply
sticking to the minimum desired decimals in that scenario.