Rework transaction signing to rely on enrichment annotations
This removes the dependency on bespoke metadata lookup, and will remove
the in-UI Ethereum provider for everything that isn't an internal dApp.

Additionally, this reuses existing resolution pathways so that the
system can consistently interpret transaction data in one place. The
existing SignType abstraction, which preceded annotations, is removed
in favor of resolving particular signing rendering components based on
the annotation type.

The biggest refactor here is that SignTransaction only takes a redirect
location for the location state. Everything else is expected to be
resolved via enrichment. This is currently only used in the Send
component, which now sends the request to redirect back to the
single-asset activity list.