Add SerialFallbackProvider for Ethers
SerialFallbackProvider is an Ethers JsonRpcProvider that is built to
fall back among a list of providers that could fulfill requests. By
default, it always uses a WebSocketProvider, but can fall back to other
providers that may or may not use WebSockets if that provider fails.

Fallbacks only kick in after exponentially backed-off retries, and
retries can, for WebSockets, reconnect the WebSocket. Additionally, the
fallback provider tracks subscriptions and reattaches them when a
provider switches or is reconnected.

The serial fallback provider will also be the place where abstractions
will occur between different types of providers, e.g. using optimized
Alchemy-specific calls for an Alchemy provider but less optimized ones
for non-Alchemy providers.