Introduce AssetDataHelper to wrap chain asset lookups
This helper wraps token-specific functionality and will handle fallbacks
in cases where Alchemy methods aren't available. Currently only does
that for token metadata.

Asset data methods take network info as a parameter and resolve the
appropriate provider by calling back into `ChainService`, and generally
log an error and return empty data if the network in question is
unsupported or fails to resolve a provider.

Currently this requires exposing `currentProvider` in
`SerialFallbackProvider` to test instanceof before calling Alchemy
utilities. Likely these will be updated separately to let non-Alchemy
providers be used, with the expectation that calls to them will be
wrapped with the asset data helper.

As part of this, add a `getMetadata` helper to the ERC20 lib, and drop
the `getBalances` helper that was in a weird halfway point between
Alchemy-specific and Tally-specific, and barely touching ERC20 details.