Match impedance between the world and LedgerService
The incoming addresses (account?) should be translated to derivation
paths when the ledgerjs library is called for signing.
The generated signature lacked all the leading 0x which is crucial to
use the already-tested libraries to handle them.
The LedgerService database is prepared to handle address resolutions
with non-normalized input addresses.