Make alarms with two time bases possible
Sometimes it would be useful to have scheduled events which run on
a two different schedules depending on external triggers. This commit
makes possible to define a base tick through the original alarm
interface (eg. 'periodInMinutes') and extend this deadline by an
integer factor.

The solution tries to enclose all the required state variables in
the base class, however the implementation in not written in
idiomatic TypeScript...