Add mainnet fork usage behind feature flag
- translate between chainID 1337<>1 on modifying transactions
- use env variables for url and chainID

Using a prodlike environment — running our code agains a mainnet fork
is valuable so we can test contract interaction and dApps integrations
that are deployed on mainnet.

For the deployed dApps to work normaly they need to see chainID 1.
Otherwise it will trigger a "sleect a supported network" response.

Hardhat mainnet fork freezes atm if the chainID is set to 1 but works
well with chainID 1337.

The soution is to translate between them when writing to the fork.