remove git dependency from the build process
Depending on the git to check version number during
build time proved to be a headache, so we have returend
to using the version from `package.json` file.

The idea is good though, probably should be placed into
the CI pipe as a step so the build itself does not depend
on the git tag.

(It caused issues with the firefox release and also internally)