inject window-provider at build time into provider bridge
The async fetch call to load the `window-provider.js` runtime
caused some really bad race conditions, so we needed a way
to have the content of the `window-provider.js` readily available
when the provider-bridge starts running.

This functionality had to work in both dev and prod environments
so the solution was to create a webpack plugin for this.

The trickiest part was finding the proper hooks in webpack
and escaping the source code so it can be stored as a string.