Correctly fix CORS issues in some Firefox versions
In other browsers, givint https://*/* permissions to content scripts
gives access to all domains irrespective of CORS requirements from
inside the extension as well. In some versions of Firefox https://*/*
must be given as an explicit permission despite having it in content
scripts.

The property to do this was already in the Firefox manifest, but it was
mistakenly under browser_specific_settings instead of at the top level.
This is now fixed.