Add check to see if we are already connected to the primary provider.
It is possible that we reconnect to the primary provider in the 15 seconds between when
attemptToReconnectToPrimaryProvider is called and the function inside of its waitAnd is executed.  We don't want to keep trying to reconnect to the primaryProvider if we are already connected to it so we break the reconnect loop in the case that we are.