Correctly show asset icons in approve asset page
The logo URL wasn't being passed along, so if we didn't have an internal
version of the logo it was never displayed. The logo URL is now used if
available, with the internal version being used as a fallback if not,
and finally our fallback logo when neither is available.