make default wallet injection more explicit
The initial flow `load window provider > message background script
for default config > if set to default overwrite politely the
window.ethereum` caused problems because window.ethereum
was not present or was occupied by others libs
when dApps started looking for it.

Changed the strategy to overwrite window.ethereum
by default and reset if tally is set not to be the default
wallet explicitly by the user.

Another solution would be for the background script
to start the communication on connect and act accordingly,
but that also has the chance for the race condition to happen.

I chose the slightly more "selfish" but more robust approach here.