accountBalance batching in indexing service
Before this change every account for every asset in every
address were updated one by one, which resulted in a lot of
small writes in bursts.

Lot of small writes in bursts are taxing on our architecture.
It slows down any persistance or communication layer.
Including indexedDB, localStorage and port communication.

Safety measures and general batching will be put in place
but the best approach is to avoid these situtions if possible.

This is step is to batch the indexedDB writes of the incoming
account balances and to prepare the batching in the UI side.