refactor updateAccountBalance to also accept arrays
So it's enough do emit a single event with all the
account balance updates instead of many.

The whole theme of this PR is also true here.
A lot of small writes in butsts slows down our
system badly. In this case it affects the port
communication which happens through webext-redux.

The ports seem to be quite sensitive to flooding
which happens when a lot of communication happens
in a very short period of time.