Extract input parsing/validation to a shared hook
In certain cases, we'll want to use validation flows separate from
specific components like SharedInput. Extracting those flows allows for
that reusability.

Notably, SharedInput had to be revised to handle undefined change
values, which is special-cased to empty strings for simple string
inputs. For parsing inputs, an empty input produces an undefined value
to indicate the blankness of the entered value.