initial store enhancer skeleton
We want to rationalize the changes coming from our store
so we get less rerenders and the port communication does
not get flooded.

To achive this the best touch point is to apply these optimisations
to the subscribe method of the redux store so any other part
benefit from the performance gain. This can be achived by
creating a store enhancer.