debounceThrottle redux subscribed listeners
All the listeners need to be debounced and not the subscribe function.
The subscription function only adds the listeners to an array.
The listeners in that array are called during every dispatch.
This means every listener separately needs to be debounceThrottled.

The debounce part is necessary so we don't flood the other parts of the
system. Throttle forces a call to be made in case the bouncing does not
stop to keep system reasonably responsive.

Note: Our first subscribe happens only when the UI is opened in any form