Spike referrals in DoggoService
The exiting ClaimService is renamed to DoggoService, contract ABI for
ClaimWithFriends is added, with a placeholder for the address (currently
unset).

The DoggoService is set up to (a) set up asset tracking for the
HuntingGround assets and (b) listen for and look up historical data on
referrals for any accounts that are being tracked. The latter are added
to the DoggoService database and can be queried. New referrals result in
an event that carries bonus total and referred user total.

No UI integration is included, and this code hasn't been tested beyond
making sure it builds.