Default typed data source to spender address
In cases where the source cannot be resolved to a known name, the source
was left undefined, which value was eventually passed to a
capitalization function that fell over on it.

We now default the source to the spender address if it could not be
resolved, which the capitalization function will leave unchanged.