Explicity type nameService cache values as possibly undefined.
When attempting to check the cache for a name that we had never resolved before - we were encountering a typeerror due to trying to access the `resolved` property and then trying to destructure a cache check that resolved to `undefined`.

The fact that querying the cache could possibly result in undefined being returned should be baked into the typing for the cache - so "undefined" was added as a possible value for both the avatar cache and the name cache - and code relying on those caches updated accordingly.