Fix isArbitraryDataSigningEnabled assignment in redux migration six.
It addresses two bugs:

The first is a bug where we were assigning `isArbitraryDatSigningEnabled` to the root LedgerState rather than assigning it to the state of each connected device.  This did not seem to have any adverse effect on the application but it did not conform to our typing.

The second is a much rarer bug where if a user was upgrading to version 0.13.0 from a pre 0.11.0 version of tally the application would become unresponsive due to pre 0.11.0 redux-states not having the "ledger" property.