Fix input spacebar bug.
When pressing spacebar in a SharedInput that was rendered as a child of the AccountItemOptionsMenu - the application would infer the spacebar as a click and close the SharedSlideUpMenu that houses the account list.  stopPropagation was innefective at stopping this behavior since the "target" of the event was the button even when spacebar was pressed while focusing a child of the button (in the former case - the SharedInput).

Changing the button to a div with a role of button allows events to capture all the way down to the clicked child element - and has the added benefit of letting us add an onKeyDown handler to explicitly handle keyboard events.