Add validation for new address names.
Added validation to ensure that users don't enter an empty address name.  Since the parseAndValidate prop on SharedInput trims values before validating them - this needs to exist inside of the AccountItemEditName component.

Also added an onPressEnter prop to SharedInput - this can easily be factored out to an onKeyDown in the future if there is a more general use case but inputs needing functionality on pressing enter is common enough that it feels like enter having its own method is OK.