Move from a single-network ChainService to many
This required breaking a few assumptions in other services. The goal
here isn't to add much additional functionality, but to take us from 1
to n networks in the background as a step toward real multi-network
support.

Includes a feature flag, MULTI_NETWORK, that when false will only
include Ethereum mainnet in the list of supported networks. When true,
it adds Arbitrum One.

Refs #7