Clear deposit input
We are saving input value in redux state to have that value saved
to perform trasactions - to save state between routes.
But when user enters the earn vault deposit input should be empty.
The best place to reset deposit input value is the main earn page
where user is selecting vaults.