Add docker build for identical outputs
Firefox requires to upload source code if
minifier is used and to be able to compile
identical output to the uploaded package.
Our builds are environment dependent at the
moment because of the minification and source
map process. Long term solution will be to upgrade
our build process to be able to produce identical
file assets, but until then we use Docker.