Update migration to annotate activities w/ timestamps
The user's preexisting activities won't have been annotated with
`blockTimestamp` by the enrichment service. Therefore, if we only remove all
blocks, the existing activities will be missing timestamps. As a solution, this
migration goes through all existing activities and annotates them with the
block timestamp before deleting the blocks.

I'm ashamed of the types here, but it's a whole can of worms partly because the
activities slice is using `createEntityAdapter`.