Rework transaction queueing to work multi-network
Rather than trying to track queues by network, this tracks a single
queue for all transactions and tracks the network inside queue entries.
When a transaction retrieval alarm goes off, the queue is traversed in
search of the first TRANSACTIONS_RETRIEVED_PER_ALARM transactions for
each network in the queue, with remaining transactions left in the queue
for future processing.

Actual transaction retrieval is factored out to its own method for
clarity.

Note that the previous approach, where each network gets its own
subproperty and list on transactionsToRetrieve, is also feasible, but it
requires extra checking and rechecking and initialization for unclear
gains.