update Dockerfile and build steps for ff
I have a hunch that because of sqlite3
install failure the postinstall script
is not run and our patches are not applied.
So just removed the remotedev-server from
the background package for build as quick
and hacky workaround until we solve the
issue properly. We don't need it for build
so it's a safe corner to cut here.

Also wanted to record the exact steps
to prepare the source and the output
for submission.