Update alchemyAssetTransferJTD to allow for erc20 and erc1155 categories.
Since calls to getAssetTransfers for Ethereum L2's do not support the "internal" and "external" categories - we add the erc20 and erc1155 categories to try to get a better picture of asset transfers.  Since we also add these categories to the L1 call to alchemy_getAssetTransfers - we need to update the jtd validator to accept erc20 and erc1155 as valid categories.