Call getLatestBaseAccountBalance as part of indexing service's handleTokenAlarm
Right now it seems that we only update the balance of the BaseAsset when indexing service detects a transfer of a base asset.  In the case that the BaseAccountBalance is not loaded correctly initially (due to network conditions) this results in the user never having their BaseAsset balance updated unless they either re-install or receive a transfer in that BaseAsset.

This change augments our 1-minute balance refresh to also refresh the BaseAccountBalance as well as token balances.