Revert for non-blocking forEach.
Per @ShadowFiend: In general I would rather let the rest of the system work on the possibility of the chain not having connected "fully" (ie with head tracking) yet, since that means it handles the possibility of being out of sync later as well. Alternatively, we could have the providers transparently handle the critical subscriptions from caller perspectives. Either way, I think we drop the await, log errors, and call it a day.