Add setSelectedNetwork action.
Since we do not yet have a concept of "selectedNetwork" in our redux store - we change the network the extension is looking at by chaging the "network" property of the currently selected address.  In the case that we're changing to an address/network combination that we have not used before - we add the addressNetwork to our list of addressNetworks to track.