Fix Polygon Gas Estimates.
Always relying on a formula that has baseFeePerGas as its primary variable is problematic on the polygon network where the baseFeePerGas is often sub 100 wei.  Using baseFeePerGas to calculate the maxFeePerGas results in a "maxFeePerGas is lower than maxPriorityFeePerGas" error pretty consistently.  This commit changes our gas fee estimations to prefer the maxFeePerGas they are passed via blocknative or the polygon gas api when those services are available.