Change error handling in get0xApiBase.
Rather than throw an error when we can't dont have an api base for a given network we return null and log that swaps are not supported on that network.  We may want to consider separating the logic of if a swap can be performed on a given network to a separate function in the future so validation is not coupled with url construction.