Change defaultWallet default value to false.
This changes the value that we initialize the application with to match the value our window provider is initialized with.  Without this change the application is in a weird state where redux AND preference service think that Tally is the default wallet but window.ethereum.isMetaMask is "false" - leading to a lot of confusion.