Rename featureFlag to isAvailable in add wallet button data
This was originally used to show or hide "add wallet" buttons based on
feature flags, but it's since modified to be based on more than just
feature flags (in particular, Ledger is displayed or not depending on
Ledger support rather than a Ledger feature flag). The updated name
for the property is decomplected from how wallet availability is
determined.