Make price test data reference checked assets
The prices data tests were explicitly expanding base assets in the
expected getPrices response, but in fact the expected behavior of
getPrices is to return those assets as-is. The test data now references
the same asset variable that is passed to getPrices in the test call,
fixing issues with the newly-introduced coinType property as well as
more clearly expressing the intended behavior of getPrices.