update network swithcer effect dependency list
Without this change the effect was running on every UI slice change,
and as a result it cleared the current state of the swap every time.
The user might have been in the middle of a swap, and then it's gone.

webext-redux patch strategy replaces the whole slice in redux
on every change that has happened to the UI slice. The effect
seems to check the reference of a dependant object. Adding
the 2 together we clear swap state on every UI slice change.

If the depency of the effect is a primitive, then useEffect is
checking for the value which is what we want.