Unify AccountSigner and SigningMethod
These two were subtly different data structures that carried roughly the
same information. Rather than keep them split, unify them as one type,
the AccountSigner, and make SigningMethod an alias for it to minimize
changes. The old AccountSigner type had an accountID for all signers;
the new structure looks more like SigningMethod in that each signing
type has its own properties to identify the signer uniquely for its
service.

SigningMethod subtypes for ledger and keyring were defined inline in
the same file; AccountSigner subtypes are instead defined in the
specific signer slices with the information needed to identify the way
to sign for an address with that signer, and then unioned into the top-
level AccountSigner type.

Minimal changes are required to the consuming code in this commit since
SigningMethod is left as an alias, and the new AccountSigner subtypes
are interface matches for the previous SigningMethod subtypes.