Remove SigningMethod
To do this, all uses of `SigningMethod` are replaced with
`AccountSigner`, and all fields named `method` or `signingMethod` are
renamed to `accountSigner`.

For completeness, the selector `selectAddressSigningMethods` becomes
`selectAccountSignersByAddress`, and `selectCurrentAccountSigningMethod`
becomes `selectCurrentAccountSigner`. `selectAccountSignersByAddress` is
lightly refactored compared to its predecessor, eliminating the need for
`selectKeyringSigningAddresses` and `selectLedgerSigningMethods` as
separate selectors.