reload the UI page if it's not rendered in 1s
In certain situations the `proxyStore.ready()` is not fired,
and in those cases a green screen is shown forever.
This effect can be reproduced with enabling network +
CPU throttling for the background script in devtools
and a couple of reload>UI open cycle. The blocked
state can be resolved if the UI thread is reloaded.

The solution waits for 1s for the script to start running and
render the first top level element. If this does not happen
then reloads the UI thread.