remove the unused custom db migrations
Dexie has its own migration solution, it's better to use that
instead of our own. (We weren't using these anyway.)

These migrations caused serious slow down (~400ms) when the
services were starting, and because of the service dependencies
this slow down propagated to the overall background startup time.

The effect become significantly worse when the user opened the
UI quickly after installation/extension reload. In those cases
the slowdown could go up to ~1500-2000ms. (Not sure why.
My best guess is that the browser was copying the db
to set up the UI thread which interefered with the migration.)

For the problem to present itself I was using a dev build,
slow 3g network throttling and 6x CPU throttling in chrome.

See #1672 debug notes + fallback fix