Move webext-redux updatesto deep-diff strategy
This strategy only sends the actual changed areas across the port, but
more importantly it only changes the objects that need to be changed in
the proxy store. This leads to less rerendering in the extension popup.

Also added are two type declarations for the diff and patch functions,
which don't have types exported by default and therefore cause
TypeScript errors under noImplicitAny. The types are declared as
restrictively as possible.