return chainId independent from permission
uniswap is using the @web3-react/metamask for every
injected provider

the @web3-react/metamask provider throws an error
if `eth_chainId` returns an error independent of the
permission for the dApp.

Promise.all(`eth_chainId`,`eth_requestAccounts` ).then ...
https://github.com/tallycash/web3-react/blob/cf822e93b3d6f8dfeef8af0a41fc176cda429f33/packages/metamask/src/index.ts#L127

And we don't have a way to change those libs, because
our PRs are not getting merged.