Fix deprecated estimated fee preview for multi-network
Really this fixes how the estimated fee is calculated in general in the
deprecated version. The reason this showed up as an issue here is because, on
Polygon, baseFeePerGas is often sub 100 wei. The incorrect calcs on this step
disregarded the max priority fee, so it resulted in 0 gwei values.