Add baseline trust to detected tokens, ignore untrusted
New tokens are only detected through enriched EVM transaction ERC20
transfer events. Currently this means any transaction we detect with
tokens will add that token, as long as the sender or recipient of the
token are tracked.

However, some scams rely on sending fake tokens that look legit to a
user. Auto-detecting these tokens and presenting them in the UI is
dangerous without flagging for the user that the tokens may be
untrusted.

This commit introduces a concept of baseline trust when deciding to
check for balances in ERC20 transfers. Baseline trust is currently
defined as a token that was SENT BY an address that is being tracked
(i.e., an account the user has added to the wallet) OR that is already
known (typically via a token list).

Currently tokens without baseline trust are ignored (i.e. they are
neither tracked nor balance-checked), but this is the core function that
will allow presenting the user with choices about these tokens.