create a memoized version of the mergeAssets
this function call is extremply expensive cpu wise
and it's also a pure function in the practical sense
so we can optimize by running it only if the input
list changes

we use a shortcut for calculating the cache key
for similarity definition. We are not checking for
all the contract addresses to be the same, but
we use the number of all the elements in the list
as an indicator that the incoming lists are the same
This is a reasonable compromise, given we are dealing
with token lists here.