Handle solely hex data to `SigningService.signData`
This function previously received non-hexed data, and the internals were
built to handle mostly non-hexed data. Adjust keyring service to use
signMessageBytes, ledger service to assume it receives hex, and signing
service to throw if it is handed non-hex data for signing.

Also align types to use HexString, even though this is not a TypeScript-
enforced distinction.